% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/est_rf.R
\name{est_rf}
\alias{est_rf}
\title{A recombination fraction function}
\usage{
est_rf(dat)
}
\arguments{
\item{dat}{an object of class \code{outcross}, \code{bc.onemap},
\code{f2.onemap}, \code{riself.onemap} or \code{risib.onemap}.}
}
\description{
This function computes the recombination fraction matrix for backcroos, F2 and outcrossing expertimental populations.
}
\examples{
dat<-simulate_pop(type="f2", n.ind = 250,
                  n.mrk = 1000, ch.len = 200,
                  dom43 = 20, dom51 = 20,
                  missing = 15, n.ch = 2,
                  verbose = TRUE)
require(onemap)
dat
r<-est_rf(dat)
}
\keyword{recombination}

