% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate_pop.R
\name{simulate_pop}
\alias{simulate_pop}
\title{A Simulation function}
\usage{
simulate_pop(type = c("bc", "f2", "out"), n.ind, n.mrk, ch.len, missing = 0,
  n.ch = 1, dom43 = 0, dom51 = 0, prob = c(1, 1, 1, 1, 1, 1, 1),
  verbose = TRUE)
}
\arguments{
\item{type}{type of cross. Must be one of the following: "bc" fro backcross, "f2" for f2 and "out" fro outcross.}

\item{n.ind}{number of individuals}

\item{n.mrk}{number of markers}

\item{ch.len}{length of chromosome in centimorgans}

\item{missing}{percentage of missing data}

\item{n.ch}{number of chromosomes}

\item{dom43}{if \code{type=="f2"}, percentage of dominant markers NOT BB, BB}

\item{dom51}{if \code{type=="f2"}, percentage of dominant markers NOT AA, AA}

\item{prob}{if \code{type=="out"}, probability of markers types A, B1, B2, B3, C, D1 and D2}

\item{verbose}{logical. If \code{TRUE}, prints the progress of the simulation per chromosome.}
}
\description{
This function simulates experimental crosses: backcroos, F2 and outcrossing populations.
}
\examples{
dat<-simulate_pop(type="f2", n.ind = 250,
                  n.mrk = 1000, ch.len = 200,
                  missing = 15, dom43 = 20,
                  dom51 = 20, n.ch = 2,
                  verbose = TRUE)
require(onemap)
dat
}

